cleanup and move to VHR tap
