--- apps/app/src/lib/createPostTypeData.ts
@@ -1,13 +1,0 @@
-import { development, LensClient } from '@lens-protocol/client'
-
-const client = new LensClient({
-  environment: development
-})
-
-const createPostTypeData = () => {
-  return client.publication.createPostTypedData
-}
-
-export default createPostTypeData
-
-export { client }
